tests/turbohtml-tests/treebuilder_coverage.dat

#data
<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//" "http://www.w3.org/TR/html4/loose.dtd">
#errors
#document
| <!DOCTYPE html "-//W3C//DTD HTML 4.01 Transitional//" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|   <body>

#data
\x00\x00\x00
#errors
#document
| <html>
|   <head>
|   <body>

#data
<head></head>\x00text
#errors
#document
| <html>
|   <head>
|   <body>
|     "text"

#data
<head></head>\x0ctext
#errors
#document
| <html>
|   <head>
|   <body>
|     "text"

#data
<math><mi>\x00text</mi></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "text"

#data
<math><mi>\x0ctext</mi></math>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "text"


#data
  text
#errors
#document
| <html>
|   <head>
|   <body>
|     "text"

#data
<table>\x00\x00</table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>

#data
<head></head></template>text
#errors
#document
| <html>
|   <head>
|   <body>
|     "text"

#data
<title></title>
#errors
#document
| <html>
|   <head>
|     <title>
|   <body>
